Refactor: category buttons 컴포넌트 분리
