+++ src/components/myPage/Collections/index.js
@@ -6,6 +6,7 @@ export { default as ItemTitle } from './ItemTitle';
 
 export { default as DateEditor } from './DateEditor';
 export { default as CategoryButton } from './CategoryButton';
+export { default as CategoryButtons } from './CategoryButtons';
 export { default as EmptyMessage } from './EmptyMessage';
 
 export { default as SkeletonWrapper } from './SkeletonWrapper';
